chg: extract script add/edit to a reusable component
