+++ frontend/src/stores/reactor.store.js
@@ -106,7 +106,13 @@ export const useReactorStore = defineStore({
         .post(`${baseUrl}/scripts/${id}/test`, { payload })
         .finally(() => (this.status.testing = false));
     },
-    async saveAndTest({ scriptId, payload, scriptPayload, testPayload }) {
+    async saveAndTest({
+      scriptId,
+      payload,
+      scriptPayload,
+      testPayload,
+      trigger,
+    }) {
       this.status.testing = true;
       try {
         let id = scriptId;
@@ -119,9 +125,13 @@ export const useReactorStore = defineStore({
           );
           id = created.id;
         }
-        const run = await fetchWrapper.post(`${baseUrl}/scripts/${id}/test`, {
-          payload: testPayload ?? payload ?? {},
-        });
+        const body = { payload: testPayload ?? payload ?? {} };
+        if (trigger?.resource_type) body.resource_type = trigger.resource_type;
+        if (trigger?.action) body.action = trigger.action;
+        const run = await fetchWrapper.post(
+          `${baseUrl}/scripts/${id}/test`,
+          body,
+        );
         const log = await fetchWrapper.get(`${baseUrl}/runs/${run.id}/log`);
         return { scriptId: id, run, log: log.log };
       } finally {
